(PC-12795)[PRO]refactor: renaming booking cancellation related email
